Move password hashing to server_base.
This is in preparation for adding the `forum_auth` option when uploading an add-on.
