+++ source_lists/campaignd
@@ -1,6 +1,5 @@
 addon/validation.cpp
 server/common/dbconn.cpp
-server/common/user_handler.cpp
 server/common/forum_user_handler.cpp
 server/common/server_base.cpp
 server/common/simple_wml.cpp
