+++ source_lists/wesnothd
@@ -1,5 +1,4 @@
 server/common/dbconn.cpp
-server/common/user_handler.cpp
 server/common/forum_user_handler.cpp
 server/common/resultsets/ban_check.cpp
 server/common/resultsets/tournaments.cpp
