+++ src/server/common/forum_user_handler.cpp
@@ -49,12 +49,12 @@ fuh::fuh(const config& c)
 	}
 }
 
-bool fuh::login(const std::string& name, const std::string& password, const std::string& seed) {
+bool fuh::login(const std::string& name, const std::string& password, const std::string& nonce) {
 	// Retrieve users' password as hash
 	std::string hash;
 
 	try {
-		hash = get_hash(name);
+		hash = get_hashed_password_from_db(name);
 	} catch (const error& e) {
 		ERR_UH << "Could not retrieve hash for user '" << name << "' :" << e.message << std::endl;
 		return false;
@@ -63,9 +63,9 @@ bool fuh::login(const std::string& name, const std::string& password, const std:
 	std::string valid_hash;
 
 	if(utils::md5::is_valid_hash(hash)) { // md5 hash
-		valid_hash = utils::md5(hash.substr(12,34), seed).base64_digest();
+		valid_hash = utils::md5(hash.substr(12,34), nonce).base64_digest();
 	} else if(utils::bcrypt::is_valid_prefix(hash)) { // bcrypt hash
-		valid_hash = utils::md5(hash, seed).base64_digest();
+		valid_hash = utils::md5(hash, nonce).base64_digest();
 	} else {
 		ERR_UH << "Invalid hash for user '" << name << "'" << std::endl;
 		return false;
@@ -86,7 +86,7 @@ std::string fuh::extract_salt(const std::string& name) {
 	std::string hash;
 
 	try {
-		hash = get_hash(name);
+		hash = get_hashed_password_from_db(name);
 	} catch (const error& e) {
 		ERR_UH << "Could not retrieve hash for user '" << name << "' :" << e.message << std::endl;
 		return "";
@@ -188,7 +188,7 @@ std::string fuh::user_info(const std::string& name) {
 	return info.str();
 }
 
-std::string fuh::get_hash(const std::string& user) {
+std::string fuh::get_hashed_password_from_db(const std::string& user) {
 	return conn_.get_user_string(db_users_table_, "user_password", user);
 }
